templates/python-starter/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.7.1
